# Remove hidden malloc calls