--- a/core/vm/eips.go
+++ b/core/vm/eips.go
@@ -278,7 +278,7 @@ func opBlobHash(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([
 func opCLZ(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([]byte, error) {
 	x := scope.Stack.pop()
 	// count leading zero bits in x
-	scope.Stack.push(new(uint256.Int).SetUint64(256 - uint64(x.BitLen())))
+	scope.Stack.push(*new(uint256.Int).SetUint64(256 - uint64(x.BitLen())))
 	return nil, nil
 }