… cmd/evm/runner.go
… core/vm/analysis_test.go
… core/vm/evm.go
… core/vm/evm_test.go
… core/vm/gas_table_test.go
… core/vm/instructions.go
… core/vm/instructions_test.go
… core/vm/runtime/runtime_test.go
… eth/tracers/internal/tracetest/calltrace_test.go
… eth/tracers/internal/tracetest/prestate_test.go
… eth/tracers/js/goja.go
… eth/tracers/js/tracer_test.go
… eth/tracers/logger/logger.go
… eth/tracers/logger/logger_test.go
… eth/tracers/native/4byte.go
… eth/tracers/native/call.go
… eth/tracers/native/mux.go
… eth/tracers/native/noop.go
… eth/tracers/tracers_test.go
… execution/consensus/consensus.go
… execution/consensus/misc/eip4844.go
… execution/exec3/calltracer/calltracer_v3.go
… polygon/aa/entry_point_tracer.go
… polygon/aa/validation_rules_tracer.go
… polygon/bor/bor.go
… polygon/tracer/bor_state_sync_txn_tracer.go
… polygon/tracer/trace_bor_state_sync_txn.go
… rpc/jsonrpc/eth_callMany.go
… rpc/jsonrpc/otterscan_trace_contract_creator.go
… rpc/jsonrpc/otterscan_trace_operations.go