This changes the argument passing pattern for uint256.Ints in the evm call sequence from pass by pointer to pass by reference.  This is to avoid the unexpected mallocs on calls due to the value escaping in the call sequence and hence being moved to the heap.

This is the performance difference from the current main:

```
markholt@Marks-MacBook-Air runtime % ~/go/bin/benchstat old.simple.txt new.simple.txt
goos: darwin
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: Intel(R) Core(TM) i5-8210Y CPU @ 1.60GHz
                                      │ old.simple.txt │            new.simple.txt            │
                                      │     sec/op     │    sec/op     vs base                │
SimpleLoop/staticcall-identity-100M-4    2376.0m ± 26%   519.7m ± 26%  -78.13% (p=0.000 n=10)
SimpleLoop/call-identity-100M-4          1139.4m ± 49%   635.3m ± 23%  -44.24% (p=0.000 n=10)
SimpleLoop/loop-100M-4                   1497.6m ± 77%   666.2m ±  2%  -55.52% (p=0.000 n=10)
SimpleLoop/loop2-100M-4                  1379.3m ± 47%   743.1m ±  5%  -46.13% (p=0.000 n=10)
SimpleLoop/call-nonexist-100M-4          1540.8m ± 36%   584.7m ± 56%  -62.05% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-4               1915.9m ± 54%   751.5m ± 36%  -60.77% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-4           1.588 ± 48%    1.051 ± 10%  -33.82% (p=0.000 n=10)
geomean                                    1.595         691.3m        -56.65%

                                      │  old.simple.txt   │             new.simple.txt             │
                                      │       B/op        │     B/op       vs base                 │
SimpleLoop/staticcall-identity-100M-4       83.61Mi ±  0%   47.03Mi ±  0%   -43.75% (p=0.000 n=10)
SimpleLoop/call-identity-100M-4             81.93Mi ±  0%   46.08Mi ±  0%   -43.75% (p=0.000 n=10)
SimpleLoop/loop-100M-4                      1.613Ki ± 32%   1.102Ki ± 25%   -31.72% (p=0.001 n=10)
SimpleLoop/loop2-100M-4                     1.336Ki ± 42%   1.102Ki ± 24%   -17.54% (p=0.000 n=10)
SimpleLoop/call-nonexist-100M-4         23321.949Ki ±  0%   1.129Ki ± 24%  -100.00% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-4              40812.334Ki ±  0%   1.129Ki ± 24%  -100.00% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-4            218.0Mi ±  0%   201.6Mi ±  0%    -7.50% (p=0.000 n=10)
geomean                                     3.106Mi         132.7Ki         -95.83%

                                      │  old.simple.txt   │            new.simple.txt             │
                                      │     allocs/op     │  allocs/op    vs base                 │
SimpleLoop/staticcall-identity-100M-4        3.425M ±  0%   2.055M ±  0%   -40.00% (p=0.000 n=10)
SimpleLoop/call-identity-100M-4              3.356M ±  0%   2.013M ±  0%   -40.00% (p=0.000 n=10)
SimpleLoop/loop-100M-4                       11.500 ± 22%    8.000 ± 12%   -30.43% (p=0.000 n=10)
SimpleLoop/loop2-100M-4                      10.000 ± 30%    8.000 ± 12%   -20.00% (p=0.000 n=10)
SimpleLoop/call-nonexist-100M-4          746262.500 ±  0%    9.000 ± 11%  -100.00% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-4              1492513.500 ±  0%    9.000 ± 11%  -100.00% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-4             4.286M ±  0%   2.857M ±  0%   -33.33% (p=0.000 n=10)
geomean                                      67.38k         1.800k         -97.33%
markholt@Marks-MacBook-Air runtime % ~/go/bin/benchstat old.call.txt new.call.txt
goos: darwin
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: Intel(R) Core(TM) i5-8210Y CPU @ 1.60GHz
       │ old.call.txt  │            new.call.txt             │
       │    sec/op     │   sec/op     vs base                │
Call-4   11.328m ± 19%   6.928m ± 5%  -38.84% (p=0.000 n=10)
```

It has the additional advantage of cleaning the perf profile tree so its easier to see other potential optimizations.